third_party/WebKit/LayoutTests/http/tests/locks/mode-shared.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Web Locks API: Shared Mode</title>
<link rel=help href="https://github.com/inexorabletash/web-locks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async t => {
  let n = 0;
  const granted = [];
  function grant(n) { return () => { granted.push(n); }; }

  await Promise.all([
    navigator.locks.acquire('a', {mode: 'shared'}, grant(++n)),
    navigator.locks.acquire('b', {mode: 'shared'}, grant(++n)),
    navigator.locks.acquire('c', {mode: 'shared'}, grant(++n)),
    navigator.locks.acquire('a', {mode: 'shared'}, grant(++n)),
    navigator.locks.acquire('b', {mode: 'shared'}, grant(++n)),
    navigator.locks.acquire('c', {mode: 'shared'}, grant(++n)),
  ]);

  assert_array_equals(granted, [1, 2, 3, 4, 5, 6]);
}, 'Lock requests are granted in order');

promise_test(async t => {
  await navigator.locks.acquire(['a', 'b'], {mode: 'shared'}, async lock => {
    assert_array_equals(lock.scope, ['a', 'b']);
    assert_equals(lock.mode, 'shared');

    // Since lock over [a, b] is held, these requests would
    // be blocked if the locks were not all 'shared'.

    await navigator.locks.acquire('a', {mode: 'shared'}, lock => {
      assert_array_equals(lock.scope, ['a']);
      assert_equals(lock.mode, 'shared');
    });

    await navigator.locks.acquire('b', {mode: 'shared'}, lock => {
      assert_array_equals(lock.scope, ['b']);
      assert_equals(lock.mode, 'shared');
    });

  });
}, 'Shared locks are not exclusive');

</script>
